+++ packages/polkahub/src/index.ts
@@ -1,4 +1,6 @@
 export {
+  AddressBalance,
+  AddressIdentity,
   AvailableAccountsContext,
   formatBalance,
   ModalContext,
@@ -10,8 +12,6 @@ export {
   useModalContext,
   usePlugin,
   usePolkaHubContext,
-  AddressBalance,
-  AddressIdentity,
 } from "@polkahub/context";
 export * from "@polkahub/ledger";
 export * from "@polkahub/mimir";
@@ -22,14 +22,22 @@ export * from "@polkahub/proxy";
 export * from "@polkahub/read-only";
 export {
   createSelectedAccountPlugin,
-  SelectAccountField,
   SelectAccount,
+  SelectAccountField,
   selectedAccountPluginId,
   useSelectedAccount,
   type SelectedAccountPlugin,
 } from "@polkahub/select-account";
 export * from "@polkahub/state";
-export * from "@polkahub/vault";
+export {
+  createPolkadotVaultProvider,
+  ManageVault,
+  polkadotVaultProviderId,
+  VaultTxModal,
+  type PolkadotVaultAccount,
+  type PolkadotVaultProvider,
+  type VaultAccountInfo,
+} from "@polkahub/vault";
 export * from "@polkahub/wallet-connect";
-export * from "./PolkaHubModal";
 export * from "./AddressInput";
+export * from "./PolkaHubModal";